requirements.txt
PyMuPDF
python-pptx
pyinstaller
